move repository interfaces to entities sub-packages
