+++ src/integrationTest/java/com/trynoice/api/subscription/SubscriptionControllerTest.java
@@ -12,8 +12,11 @@ import com.stripe.model.SubscriptionItemCollection;
 import com.stripe.model.checkout.Session;
 import com.trynoice.api.identity.entities.AuthUser;
 import com.trynoice.api.subscription.entities.Customer;
+import com.trynoice.api.subscription.entities.CustomerRepository;
 import com.trynoice.api.subscription.entities.Subscription;
 import com.trynoice.api.subscription.entities.SubscriptionPlan;
+import com.trynoice.api.subscription.entities.SubscriptionPlanRepository;
+import com.trynoice.api.subscription.entities.SubscriptionRepository;
 import com.trynoice.api.subscription.models.SubscriptionFlowParams;
 import com.trynoice.api.subscription.models.SubscriptionPlanView;
 import com.trynoice.api.testing.AuthTestUtils;
