+++ src/main/java/com/trynoice/api/identity/AccountService.java
@@ -7,7 +7,9 @@ import com.auth0.jwt.exceptions.JWTVerificationException;
 import com.github.benmanes.caffeine.cache.Cache;
 import com.trynoice.api.contracts.AccountServiceContract;
 import com.trynoice.api.identity.entities.AuthUser;
+import com.trynoice.api.identity.entities.AuthUserRepository;
 import com.trynoice.api.identity.entities.RefreshToken;
+import com.trynoice.api.identity.entities.RefreshTokenRepository;
 import com.trynoice.api.identity.exceptions.AccountNotFoundException;
 import com.trynoice.api.identity.exceptions.BearerJwtAuthenticationException;
 import com.trynoice.api.identity.exceptions.DuplicateEmailException;
