+++ src/main/java/com/trynoice/api/subscription/entities/SubscriptionRepository.java
@@ -1,7 +1,6 @@
-package com.trynoice.api.subscription;
+package com.trynoice.api.subscription.entities;
 
 import com.trynoice.api.platform.BasicEntityRepository;
-import com.trynoice.api.subscription.entities.Subscription;
 import lombok.NonNull;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -17,7 +16,7 @@ import java.util.Optional;
  * com.trynoice.api.platform.BasicEntity BasicEntity}.
  */
 @Repository
-interface SubscriptionRepository extends BasicEntityRepository<Subscription, Long> {
+public interface SubscriptionRepository extends BasicEntityRepository<Subscription, Long> {
 
     /**
      * Find a {@link Subscription} entity by its provider assigned subscription id.
