+++ src/test/java/com/trynoice/api/identity/AccountServiceTest.java
@@ -4,7 +4,9 @@ import com.auth0.jwt.JWT;
 import com.auth0.jwt.algorithms.Algorithm;
 import com.github.benmanes.caffeine.cache.Cache;
 import com.trynoice.api.identity.entities.AuthUser;
+import com.trynoice.api.identity.entities.AuthUserRepository;
 import com.trynoice.api.identity.entities.RefreshToken;
+import com.trynoice.api.identity.entities.RefreshTokenRepository;
 import com.trynoice.api.identity.exceptions.AccountNotFoundException;
 import com.trynoice.api.identity.exceptions.DuplicateEmailException;
 import com.trynoice.api.identity.exceptions.RefreshTokenVerificationException;
