+++ src/test/java/com/trynoice/api/sound/SoundServiceTest.java
@@ -1,6 +1,7 @@
 package com.trynoice.api.sound;
 
 import com.trynoice.api.contracts.SubscriptionServiceContract;
+import com.trynoice.api.sound.entities.LibraryManifestRepository;
 import com.trynoice.api.sound.exceptions.SegmentAccessDeniedException;
 import lombok.NonNull;
 import lombok.val;
