+++ src/test/java/com/trynoice/api/subscription/SubscriptionServiceTest.java
@@ -6,8 +6,11 @@ import com.stripe.exception.StripeException;
 import com.stripe.model.checkout.Session;
 import com.trynoice.api.contracts.AccountServiceContract;
 import com.trynoice.api.subscription.entities.Customer;
+import com.trynoice.api.subscription.entities.CustomerRepository;
 import com.trynoice.api.subscription.entities.Subscription;
 import com.trynoice.api.subscription.entities.SubscriptionPlan;
+import com.trynoice.api.subscription.entities.SubscriptionPlanRepository;
+import com.trynoice.api.subscription.entities.SubscriptionRepository;
 import com.trynoice.api.subscription.exceptions.DuplicateSubscriptionException;
 import com.trynoice.api.subscription.exceptions.SubscriptionNotFoundException;
 import com.trynoice.api.subscription.exceptions.SubscriptionPlanNotFoundException;
